+++ src/ui/channel_navigation_mode_renderer.h
@@ -1,0 +1,11 @@
+/**
+ * @file channel_navigation_mode_renderer.h
+ * @brief Companion translation-unit boundary for ChannelNavigationRenderer mode selection.
+ * @author Axel Napolitano
+ * @copyright 2026 Axel Napolitano
+ * @license PolyForm-Noncommercial-1.0.0
+ */
+
+#pragma once
+
+#include "ui/channel_navigation_renderer.h"